+++ src/Muralis.App/Views/FavoritesPage.xaml.cs
@@ -1,3 +1,4 @@
+using Microsoft.UI.Xaml;
 using Microsoft.UI.Xaml.Controls;
 using Muralis.App.ViewModels;
 using Muralis.Core.Models;
@@ -10,9 +11,12 @@ public sealed partial class FavoritesPage : Page
     {
         ViewModel = App.GetService<FavoritesViewModel>();
         InitializeComponent();
+        Unloaded += OnUnloaded;
     }
 
     public FavoritesViewModel ViewModel { get; }
+
+    private void OnUnloaded(object sender, RoutedEventArgs e) => ViewModel.DetachFromPage();
 
     private void OnWallpaperItemClick(object sender, ItemClickEventArgs e) =>
         ViewModel.OpenWallpaperCommand.Execute(e.ClickedItem as Wallpaper);